+++ .settings/com.palantir.typescript.prefs
@@ -1,5 +1,5 @@
 build.path.exportedFolder=
 build.path.sourceFolder=ts/src/ts
 compiler.outputDirOption=ts/build/ts
-compiler.outputFileOption=
+compiler.outputFileOption=QRCode.js
 eclipse.preferences.version=1
