+++ ts/src/ts/qrcode/ErrorCorrectLevel.ts
@@ -1,0 +1,32 @@
+'use strict';
+namespace qrcode {
+
+  /**
+   * ErrorCorrectLevel
+   * @author Kazuhiko Arase
+   */
+  export class ErrorCorrectLevel {
+
+    /**
+     * 7%
+     */
+    public static L = 1;
+
+    /**
+     * 15%
+     */
+    public static M = 0;
+
+    /**
+     * 25%
+     */
+    public static Q = 3;
+
+    /**
+     * 30%
+     */
+    public static H = 2;
+  }
+
+
+}
