+++ ts/src/ts/qrcode/MaskPattern.ts
@@ -1,0 +1,50 @@
+'use strict';
+namespace qrcode {
+
+  /**
+   * MaskPattern
+   * @author Kazuhiko Arase
+   */
+  export class MaskPattern {
+
+    /**
+     * mask pattern 000
+     */
+    public static PATTERN000 = 0;
+
+    /**
+     * mask pattern 001
+     */
+    public static PATTERN001 = 1;
+
+    /**
+     * mask pattern 010
+     */
+    public static PATTERN010 = 2;
+
+    /**
+     * mask pattern 011
+     */
+    public static PATTERN011 = 3;
+
+    /**
+     * mask pattern 100
+     */
+    public static PATTERN100 = 4;
+
+    /**
+     * mask pattern 101
+     */
+    public static PATTERN101 = 5;
+
+    /**
+     * mask pattern 110
+     */
+    public static PATTERN110 = 6;
+
+    /**
+     * mask pattern 111
+     */
+    public static PATTERN111 = 7;
+  }
+}
