+++ ts/src/ts/qrcode/io/InputStream.ts
@@ -1,0 +1,14 @@
+'use strict';
+namespace qrcode.io {
+
+  /**
+   * InputStream
+   * @author Kazuhiko Arase
+   */
+  export abstract class InputStream {
+    constructor() {}
+    public abstract readByte() : number;
+    public close() : void {
+    }
+  }
+}
