+++ ts/src/ts/qrcode/io/OutputStream.ts
@@ -1,0 +1,22 @@
+'use strict';
+namespace qrcode.io {
+
+  /**
+   * OutputStream
+   * @author Kazuhiko Arase
+   */
+  export abstract class OutputStream {
+    constructor() {}
+    public abstract writeByte(b : number) : void;
+    public writeBytes(bytes : number[]) : void {
+      for (var i = 0; i < bytes.length; i += 1) {
+        this.writeByte(bytes[i]);
+      }
+    }
+    public flush() : void {
+    }
+    public close() : void {
+      this.flush();
+    }
+  }
+}
