+++ ts/src/ts/test.html
@@ -3,7 +3,6 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html;charset=UTF-8" />
 <script src="../../build/ts/QRCode.js"></script>
-<script src="../../build/ts/test.js"></script>
 </head>
 <body style="padding:16px;" >
 </body>
